mocktikv: move coprHandler to standalone file (#24067)
