+++ store/mockstore/mocktikv/rpc.go
@@ -33,7 +33,6 @@ import (
 	"github.com/pingcap/parser/terror"
 	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/store/tikv/tikvrpc"
-	"github.com/pingcap/tipb/go-tipb"
 )
 
 // For gofail injection.
@@ -551,48 +550,6 @@ func (h kvHandler) handleSplitRegion(req *kvrpcpb.SplitRegionRequest) *kvrpcpb.S
 		resp.Regions = append(resp.Regions, newRegion)
 	}
 	return resp
-}
-
-func drainRowsFromExecutor(ctx context.Context, e executor, req *tipb.DAGRequest) (tipb.Chunk, error) {
-	var chunk tipb.Chunk
-	for {
-		row, err := e.Next(ctx)
-		if err != nil {
-			return chunk, errors.Trace(err)
-		}
-		if row == nil {
-			return chunk, nil
-		}
-		for _, offset := range req.OutputOffsets {
-			chunk.RowsData = append(chunk.RowsData, row[offset]...)
-		}
-	}
-}
-
-type coprHandler struct {
-	*Session
-}
-
-func (h coprHandler) handleBatchCopRequest(ctx context.Context, req *coprocessor.BatchRequest) (*mockBatchCopDataClient, error) {
-	client := &mockBatchCopDataClient{}
-	for _, ri := range req.Regions {
-		cop := coprocessor.Request{
-			Tp:      kv.ReqTypeDAG,
-			Data:    req.Data,
-			StartTs: req.StartTs,
-			Ranges:  ri.Ranges,
-		}
-		_, exec, dagReq, err := h.buildDAGExecutor(&cop)
-		if err != nil {
-			return nil, errors.Trace(err)
-		}
-		chunk, err := drainRowsFromExecutor(ctx, exec, dagReq)
-		if err != nil {
-			return nil, errors.Trace(err)
-		}
-		client.chunks = append(client.chunks, chunk)
-	}
-	return client, nil
 }
 
 // Client is a client that sends RPC.
